requirements.txt
maxapi-python
httpx
PyYAML
